days_between functionality
